Move users client to users-specific base test
