+++ tests/Integration/AbstractIntegrationTestCase.php
@@ -7,7 +7,6 @@ namespace SmartAssert\ApiClient\Tests\Integration;
 use GuzzleHttp\Client as HttpClient;
 use GuzzleHttp\Psr7\HttpFactory;
 use PHPUnit\Framework\TestCase;
-use SmartAssert\ApiClient\UsersClient;
 use SmartAssert\ServiceClient\Client as ServiceClient;
 use SmartAssert\ServiceClient\ExceptionFactory\CurlExceptionFactory;
 use SmartAssert\ServiceClient\ResponseFactory\ResponseFactory;
@@ -19,17 +18,7 @@ abstract class AbstractIntegrationTestCase extends TestCase
     protected const USER2_EMAIL = 'user1@example.com';
     protected const USER2_PASSWORD = 'password';
 
-    protected static UsersClient $client;
-
-    public static function setUpBeforeClass(): void
-    {
-        self::$client = new UsersClient(
-            'http://localhost:9081',
-            self::createServiceClient(),
-        );
-    }
-
-    private static function createServiceClient(): ServiceClient
+    protected static function createServiceClient(): ServiceClient
     {
         $httpFactory = new HttpFactory();
