+++ tests/Integration/User/AbstractUserTestCase.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SmartAssert\ApiClient\Tests\Integration\User;
+
+use SmartAssert\ApiClient\Tests\Integration\AbstractIntegrationTestCase;
+use SmartAssert\ApiClient\UsersClient;
+
+abstract class AbstractUserTestCase extends AbstractIntegrationTestCase
+{
+    protected static UsersClient $client;
+
+    public static function setUpBeforeClass(): void
+    {
+        self::$client = new UsersClient(
+            'http://localhost:9081',
+            self::createServiceClient(),
+        );
+    }
+}
