+++ tests/Integration/User/CreateTest.php
@@ -5,10 +5,9 @@ declare(strict_types=1);
 namespace SmartAssert\ApiClient\Tests\Integration\User;
 
 use SmartAssert\ApiClient\Exception\UserAlreadyExistsException;
-use SmartAssert\ApiClient\Tests\Integration\AbstractIntegrationTestCase;
 use SmartAssert\ServiceClient\Exception\UnauthorizedException;
 
-class CreateTest extends AbstractIntegrationTestCase
+class CreateTest extends AbstractUserTestCase
 {
     public function testCreateUnauthorized(): void
     {
