+++ tests/Integration/User/GetApiKeyTest.php
@@ -5,10 +5,9 @@ declare(strict_types=1);
 namespace SmartAssert\ApiClient\Tests\Integration\User;
 
 use SmartAssert\ApiClient\Model\ApiKey;
-use SmartAssert\ApiClient\Tests\Integration\AbstractIntegrationTestCase;
 use SmartAssert\ServiceClient\Exception\UnauthorizedException;
 
-class GetApiKeyTest extends AbstractIntegrationTestCase
+class GetApiKeyTest extends AbstractUserTestCase
 {
     public function testGetUserApiKeyInvalidToken(): void
     {
