+++ tests/Integration/User/RefreshTokenTest.php
@@ -5,10 +5,9 @@ declare(strict_types=1);
 namespace SmartAssert\ApiClient\Tests\Integration\User;
 
 use SmartAssert\ApiClient\Model\RefreshableToken;
-use SmartAssert\ApiClient\Tests\Integration\AbstractIntegrationTestCase;
 use SmartAssert\ServiceClient\Exception\UnauthorizedException;
 
-class RefreshTokenTest extends AbstractIntegrationTestCase
+class RefreshTokenTest extends AbstractUserTestCase
 {
     public function testRefreshInvalidRefreshToken(): void
     {
