+++ tests/Integration/User/VerifyTokenTest.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace SmartAssert\ApiClient\Tests\Integration\User;
 
-use SmartAssert\ApiClient\Tests\Integration\AbstractIntegrationTestCase;
-
-class VerifyTokenTest extends AbstractIntegrationTestCase
+class VerifyTokenTest extends AbstractUserTestCase
 {
     public function testVerifySuccess(): void
     {
